added POST (working) method for creation, PUT (still to work) method for modification. ONTO making PUT requests work

+++ src/main/java/dto/JourneyDTO.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 
 @Path("journey")
 public class JourneyDTO implements Serializable {
-    private long id, idCitizen, idVehicle;
+    private int id, idCitizen, idVehicle;
     private int departure;
     private int arrival;
     private int commute;
